Refactor & Fix param
