+++ selfdrive/ui/sunnypilot/qt/offroad/settings/lateral/max_time_offroad.h
@@ -1,0 +1,23 @@
+/**
+ * Copyright (c) 2021-, Haibin Wen, sunnypilot, and a number of other contributors.
+ *
+ * This file is part of sunnypilot and is licensed under the MIT License.
+ * See the LICENSE.md file in the root directory for more details.
+ */
+
+#pragma once
+
+#include "selfdrive/ui/sunnypilot/ui.h"
+#include "selfdrive/ui/sunnypilot/qt/offroad/settings/settings.h"
+#include "selfdrive/ui/sunnypilot/qt/widgets/controls.h"
+
+class MaxTimeOffroad : public OptionControlSP {
+  Q_OBJECT
+
+public:
+  MaxTimeOffroad();
+  void refresh();
+
+private:
+  Params params;
+};
